Unify PackUnitCell
