+++ client/packages/invoices/src/InboundShipment/DetailView/ContentArea/columns.ts
@@ -13,8 +13,7 @@ import {
 } from '@openmsupply-client/common';
 import {
   LocationRowFragment,
-  PackUnitMultipleCell,
-  getPackUnitCell,
+  PackUnitCell,
   useInitUnitStore,
 } from '@openmsupply-client/system';
 import { InboundItem } from './../../../types';
@@ -239,7 +238,7 @@ export const useInboundShipmentColumns = () => {
         key: 'packUnit',
         label: 'label.pack',
         sortable: false,
-        Cell: PackUnitMultipleCell({
+        Cell: PackUnitCell({
           getItemId: row => {
             if ('lines' in row) return '';
             else return row?.item?.id;
@@ -301,10 +300,10 @@ export const useExpansionColumns = (): Column<InboundLineFragment>[] =>
       key: 'packUnit',
       label: 'label.pack',
       sortable: false,
-      Cell: getPackUnitCell({
+      Cell: PackUnitCell({
         getItemId: row => row?.item.id,
-        getPackSize: row => {
-          return row?.packSize || 1;
+        getPackSizes: row => {
+          return [row?.packSize ?? 1];
         },
         getUnitName: row => row?.item.unitName ?? null,
       }),
