+++ client/packages/system/src/Item/Components/ItemVariant/index.ts
@@ -2,4 +2,4 @@ export * from './PackUnitCell';
 export * from './PackUnitSelectCell';
 export * from './PackUnitQuantityCell';
 export * from './PackUnitEntryCell';
-export * from './PackUnitMultipleCell';
+export * from './PackUnitCell';
